added full POST registration and login windows, refactored navbar and userStore

+++ metasnacks/server/models/models.js
@@ -5,7 +5,7 @@ const User = sequelize.define('user',{
     id:{type: DataTypes.INTEGER, primaryKey: true, autoIncrement: true},
     email:{type: DataTypes.STRING, unique:true, allowNull:false},
     organisation_name:{type: DataTypes.STRING, allowNull:false},
-    itn:{type: DataTypes.INTEGER, unique:true, allowNull:false},
+    itn:{type: DataTypes.BIGINT, unique:true, allowNull:false},
     password: {type: DataTypes.STRING, allowNull:false},
     role: {type: DataTypes.STRING, defaultValue: "USER"}
 })
